Refactor codebase: All files now under 100 lines
Major refactoring to improve code maintainability and readability by breaking down large files into focused, single-responsibility modules.

## API Layer Refactoring (src/api/)
- Broke down api.ts (305 lines) into modular structure:
  - client.ts (6 lines) - OpenAI client setup
  - prompts.ts (92 lines) - System prompts
  - aiResponse.ts (71 lines) - AI response generation
  - toolConfig.ts (73 lines) - Tool config generation
  - types.ts (13 lines) - API types
  - index.ts (3 lines) - Public API exports
- Each file focused on single responsibility
- Cleaner imports and better testability

## Tool Templates Refactoring (src/templates/)
- Broke down toolTemplates.ts (400+ lines) into categories:
  - calculator.ts (98 lines) - Calculator templates
  - converter.ts (74 lines) - Unit converter templates
  - generator.ts (88 lines) - Password generator templates
  - analyzer.ts (73 lines) - BMI calculator template
  - index.ts (28 lines) - Template registry and exports
- Organized by tool type for easier maintenance
- Easier to add new templates

## Tool Components Refactoring (src/components/tools/)
- Broke down ToolInput.tsx (160 lines) into specialized inputs:
  - inputs/BaseInput.tsx (20 lines) - Reusable wrapper
  - inputs/TextNumberInput.tsx (61 lines) - Text & number inputs
  - inputs/SelectTextareaInput.tsx (63 lines) - Select & textarea
  - inputs/SliderCheckboxInput.tsx (60 lines) - Slider & checkbox
  - inputs/DateColorInput.tsx (66 lines) - Date & color pickers
  - ToolInput.tsx (35 lines) - Dispatcher/router

- Broke down ToolOutput.tsx (150 lines) into output types:
  - outputs/BaseOutput.tsx (55 lines) - Reusable wrapper with copy
  - outputs/TextNumberOutput.tsx (44 lines) - Text & number display
  - outputs/CardCodeOutput.tsx (48 lines) - Card & code blocks
  - outputs/ListTableOutput.tsx (77 lines) - Lists & tables
  - ToolOutput.tsx (28 lines) - Dispatcher/router

## Benefits
✅ All files under 100 lines (easier to navigate)
✅ Single responsibility per file (easier to understand)
✅ Better code organization and discoverability
✅ Easier to test individual modules
✅ Reduced cognitive load when reading code
✅ Cleaner git diffs for future changes
✅ Follows industry best practices

## Files Removed
- src/api.ts → src/api/ (6 files)
- src/utils/toolTemplates.ts → src/templates/ (5 files)

## Testing
- ✅ Build successful
- ✅ TypeScript compilation passes
- ✅ All imports updated correctly
- ✅ No breaking changes to public API
