+++ src/api/client.ts
@@ -1,0 +1,6 @@
+import OpenAI from "openai";
+
+export const client = new OpenAI({
+  apiKey: import.meta.env.VITE_OPENAI_API_KEY,
+  dangerouslyAllowBrowser: true,
+});
